+++ .nowignore
@@ -1,5 +1,9 @@
 node_modules
-.env
-frontend
 README.md
 schema.txt
+src
+pages
+.babelrc
+global.d.ts
+next.config.js
+tsconfig.json
