+++ backend/db/tag.js
@@ -1,0 +1,9 @@
+import { Schema, model } from "mongoose";
+
+const TagSchema = new Schema({
+  
+});
+
+const Tag = model('Tag', TagSchema);
+
+export default Tag
