+++ backend/db/team.js
@@ -1,0 +1,9 @@
+import { Schema, model } from "mongoose";
+
+const TeamSchema = new Schema({
+  
+});
+
+const Team = model('Team', TeamSchema);
+
+export default User
