+++ backend/db/user.js
@@ -1,0 +1,9 @@
+import { Schema, model } from "mongoose";
+
+const UserSchema = new Schema({
+  
+});
+
+const User = model('User', UserSchema);
+
+export default User
