+++ backend/resolvers/index.js
@@ -1,0 +1,9 @@
+const resolvers = {
+  Query: {
+    allUsers() {
+      return [{Id: 1, FirstName: "asdasd", LastName: "asidoasdas" }];
+    }
+  }
+};
+
+export default resolvers;
