+++ frontend/tsconfig.json
@@ -15,5 +15,10 @@
     "sourceMap": true,
     "strict": true,
     "target": "esnext"
-  }
+  },
+  "exclude": [
+    "node_modules",
+    "**/*.spec.ts",
+    "backend"
+  ]
 }
